requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Session==0.5.0
mysql-connector-python==8.2.0
bcrypt==4.1.1
reportlab==4.0.7
Pillow==10.1.0
Flask-Mail==0.10.0
python-dotenv==1.0.0
Werkzeug==3.0.1



